+++ ndc-lambda-sdk/src/schema.ts
@@ -15,7 +15,7 @@ export type FunctionDefinition = {
   ndcKind: FunctionNdcKind
   description: string | null,
   arguments: ArgumentDefinition[] // Function arguments are ordered
-  resultType: TypeDefinition,
+  resultType: TypeReference,
   parallelDegree: number | null,
 }
 
@@ -27,7 +27,7 @@ export enum FunctionNdcKind {
 export type ArgumentDefinition = {
   argumentName: string,
   description: string | null,
-  type: TypeDefinition
+  type: TypeReference
 }
 
 export type ObjectTypeDefinitions = {
@@ -36,87 +36,97 @@ export type ObjectTypeDefinitions = {
 
 export type ObjectTypeDefinition = {
   description: string | null,
-  properties: ObjectPropertyDefinition[]
+  properties: ObjectPropertyDefinition[],
+  isRelaxedType: boolean
 }
 
 export type ObjectPropertyDefinition = {
   propertyName: string,
   description: string | null,
-  type: TypeDefinition,
+  type: TypeReference,
 }
 
 export type ScalarTypeDefinitions = {
   [scalarTypeName: string]: ScalarTypeDefinition
 }
 
-export type ScalarTypeDefinition = Record<string, never> // Empty object, for now
+export type ScalarTypeDefinition = BuiltInScalarTypeDefinition | RelaxedScalarTypeDefinition
 
-export type TypeDefinition = ArrayTypeDefinition | NullableTypeDefinition | NamedTypeDefinition
-
-export type ArrayTypeDefinition = {
-  type: "array"
-  elementType: TypeDefinition
+export type BuiltInScalarTypeDefinition = {
+  type: "built-in",
 }
 
-export type NullableTypeDefinition = {
+export type RelaxedScalarTypeDefinition = {
+  type: "relaxed-type"
+  usedIn: TypePathSegment[][]
+}
+
+export type TypeReference = ArrayTypeReference | NullableTypeReference | NamedTypeReference
+
+export type ArrayTypeReference = {
+  type: "array"
+  elementType: TypeReference
+}
+
+export type NullableTypeReference = {
   type: "nullable",
   nullOrUndefinability: NullOrUndefinability
-  underlyingType: TypeDefinition
+  underlyingType: TypeReference
 }
 
-export type NamedTypeDefinition = NamedObjectTypeDefinition | NamedScalarTypeDefinition
+export type NamedTypeReference = NamedObjectTypeReference | NamedScalarTypeReference
 
-export type NamedObjectTypeDefinition = {
+export type NamedObjectTypeReference = {
   type: "named"
   name: string
   kind: "object"
 }
 
-export type NamedScalarTypeDefinition = CustomNamedScalarTypeDefinition | BuiltInScalarTypeDefinition
+export type NamedScalarTypeReference = CustomNamedScalarTypeReference | BuiltInScalarTypeReference
 
-export type BuiltInScalarTypeDefinition = StringScalarTypeDefinition | FloatScalarTypeDefinition | BooleanScalarTypeDefinition | BigIntScalarTypeDefinition | DateTimeScalarTypeDefinition | JSONScalarTypeDefinition
+export type BuiltInScalarTypeReference = StringScalarTypeReference | FloatScalarTypeReference | BooleanScalarTypeReference | BigIntScalarTypeReference | DateTimeScalarTypeReference | JSONScalarTypeReference
 
-export type CustomNamedScalarTypeDefinition = {
+export type CustomNamedScalarTypeReference = {
   type: "named"
   name: string
   kind: "scalar"
 }
 
-export type StringScalarTypeDefinition = {
+export type StringScalarTypeReference = {
   type: "named"
   name: BuiltInScalarTypeName.String
   kind: "scalar"
   literalValue?: string
 }
 
-export type FloatScalarTypeDefinition = {
+export type FloatScalarTypeReference = {
   type: "named"
   name: BuiltInScalarTypeName.Float
   kind: "scalar"
   literalValue?: number
 }
 
-export type BooleanScalarTypeDefinition = {
+export type BooleanScalarTypeReference = {
   type: "named"
   name: BuiltInScalarTypeName.Boolean
   kind: "scalar"
   literalValue?: boolean
 }
 
-export type BigIntScalarTypeDefinition = {
+export type BigIntScalarTypeReference = {
   type: "named"
   name: BuiltInScalarTypeName.BigInt
   kind: "scalar"
   literalValue?: bigint
 }
 
-export type DateTimeScalarTypeDefinition = {
+export type DateTimeScalarTypeReference = {
   type: "named"
   name: BuiltInScalarTypeName.DateTime
   kind: "scalar"
 }
 
-export type JSONScalarTypeDefinition = {
+export type JSONScalarTypeReference = {
   type: "named"
   name: BuiltInScalarTypeName.JSON
   kind: "scalar"
@@ -124,7 +134,7 @@ export type JSONScalarTypeDefinition = {
 
 // If there are compiler errors on this function, ensure that BuiltInScalarTypeDefinition has a type in
 // its union for every BuiltInScalarTypeName enum member, and vice versa.
-function builtInScalarTypeAssertionTest(a: BuiltInScalarTypeDefinition["name"], b: BuiltInScalarTypeName): void {
+function builtInScalarTypeAssertionTest(a: BuiltInScalarTypeReference["name"], b: BuiltInScalarTypeName): void {
   a = b;
   b = a;
 }
@@ -190,6 +200,32 @@ export class JSONValue {
   }
 }
 
+export type TypePathSegment =
+  { segmentType: "FunctionParameter", functionName: string, parameterName: string }
+  | { segmentType: "FunctionReturn", functionName: string }
+  | { segmentType: "ObjectProperty", typeName: string, propertyName: string }
+  | { segmentType: "Array" }
+  | { segmentType: "TypeParameter", typeName: string, index: number }
+  | { segmentType: "IndexSignature", typeName: string, sigIndex: number, component: "key" | "value" }
+  | { segmentType: "UnionMember", typeName: string, memberIndex: number }
+
+export function typePathToString(segments: TypePathSegment[]): string {
+  return segments.map(typePathSegmentToString).join(", ")
+}
+
+export function typePathSegmentToString(segment: TypePathSegment): string {
+  switch (segment.segmentType) {
+    case "FunctionParameter": return `function '${segment.functionName}' parameter '${segment.parameterName}'`
+    case "FunctionReturn": return `function '${segment.functionName}' return value`
+    case "ObjectProperty": return `type '${segment.typeName}' property '${segment.propertyName}'`
+    case "Array": return `array type`
+    case "TypeParameter": return `type '${segment.typeName}' type parameter index '${segment.index}'`
+    case "IndexSignature": return `type '${segment.typeName}' type signature index '${segment.sigIndex}' ${segment.component} type`
+    case "UnionMember": return `type '${segment.typeName}' union member index '${segment.memberIndex}'`
+    default: return unreachable(segment["segmentType"])
+  }
+}
+
 export function getNdcSchema(functionsSchema: FunctionsSchema): sdk.SchemaResponse {
   const functions = Object.entries(functionsSchema.functions);
 
@@ -197,7 +233,7 @@ export function getNdcSchema(functionsSchema: FunctionsSchema): sdk.SchemaRespon
     return {
       fields: Object.fromEntries(objDef.properties.map(propDef => {
         const objField: sdk.ObjectField = {
-          type: convertTypeDefinitionToSdkType(propDef.type),
+          type: convertTypeReferenceToSdkType(propDef.type),
           ...(propDef.description ? { description: propDef.description } : {})
         }
         return [propDef.propertyName, objField];
@@ -226,12 +262,12 @@ export function getNdcSchema(functionsSchema: FunctionsSchema): sdk.SchemaRespon
   }
 }
 
-function convertTypeDefinitionToSdkType(typeDef: TypeDefinition): sdk.Type {
-  switch (typeDef.type) {
-    case "array": return { type: "array", element_type: convertTypeDefinitionToSdkType(typeDef.elementType) }
-    case "nullable": return { type: "nullable", underlying_type: convertTypeDefinitionToSdkType(typeDef.underlyingType) }
-    case "named": return { type: "named", name: typeDef.name }
-    default: return unreachable(typeDef["type"])
+function convertTypeReferenceToSdkType(typeRef: TypeReference): sdk.Type {
+  switch (typeRef.type) {
+    case "array": return { type: "array", element_type: convertTypeReferenceToSdkType(typeRef.elementType) }
+    case "nullable": return { type: "nullable", underlying_type: convertTypeReferenceToSdkType(typeRef.underlyingType) }
+    case "named": return { type: "named", name: typeRef.name }
+    default: return unreachable(typeRef["type"])
   }
 }
 
@@ -241,7 +277,7 @@ function convertFunctionDefinitionToSdkSchemaType(function_name: string, definit
       .map(argDef =>
         [ argDef.argumentName,
           {
-            type: convertTypeDefinitionToSdkType(argDef.type),
+            type: convertTypeReferenceToSdkType(argDef.type),
             ...(argDef.description ? { description: argDef.description } : {}),
           }
         ]
@@ -250,7 +286,7 @@ function convertFunctionDefinitionToSdkSchemaType(function_name: string, definit
   return {
     name: function_name,
     arguments: Object.fromEntries(args),
-    result_type: convertTypeDefinitionToSdkType(definition.resultType),
+    result_type: convertTypeReferenceToSdkType(definition.resultType),
     ...(definition.description ? { description: definition.description } : {}),
   }
 }
@@ -268,10 +304,29 @@ export function printSchemaListing(functionNdcKind: FunctionNdcKind, functionDef
   }
 }
 
+export function printRelaxedTypesWarning(functionsSchema: FunctionsSchema) {
+  const relaxedTypes = Object.entries(functionsSchema.scalarTypes).flatMap<[string, RelaxedScalarTypeDefinition]>(([scalarTypeName, definition]) => {
+    return definition.type === "relaxed-type"
+      ? [[scalarTypeName, definition]]
+      : []
+  });
+  if (relaxedTypes.length > 0) {
+    console.error("The following unsupported types have been defined as custom scalar types with no runtime validation ('relaxed types').")
+    console.error("Use of relaxed types may cause runtime issues if invalid values are passed for these types.")
+    console.error("It is recommended that you replace them with equivalent supported types so that your schema can be properly defined and validated, then remove the @allowrelaxedtypes tag from the functions involved.")
+    for (const [scalarTypeName, definition] of relaxedTypes) {
+      console.error(`  * '${scalarTypeName}' - used in:`);
+      for (const usedIn of definition.usedIn) {
+        console.error(`    * ${typePathToString(usedIn)}`);
+      }
+    }
+  }
+}
+
 export function isTypeNameBuiltInScalar(typeName: string): typeName is BuiltInScalarTypeName {
   return Object.values(BuiltInScalarTypeName).find(builtInScalarTypeName => typeName === builtInScalarTypeName) !== undefined;
 }
 
-export function isBuiltInScalarTypeDefinition(typeDefinition: NamedScalarTypeDefinition): typeDefinition is BuiltInScalarTypeDefinition {
-  return isTypeNameBuiltInScalar(typeDefinition.name);
+export function isBuiltInScalarTypeReference(typeReference: NamedScalarTypeReference): typeReference is BuiltInScalarTypeReference {
+  return isTypeNameBuiltInScalar(typeReference.name);
 }