+++ ndc-lambda-sdk/test/inference/relaxed-types/union-types-invalid.ts
@@ -1,0 +1,10 @@
+type AliasedUnion = string | void;
+
+/** @allowrelaxedtypes */
+export function unionTypes(
+  numberOrString: number | Promise<string>,
+  aliasedUnion: AliasedUnion,
+  unionedObjects: { prop1: never } | { prop2: string }
+): string {
+  return ""
+}