+++ ndc-lambda-sdk/test/inference/relaxed-types/union-types-valid.ts
@@ -1,0 +1,10 @@
+type AliasedUnion = string | boolean;
+
+/** @allowrelaxedtypes */
+export function unionTypes(
+  numberOrString: number | string,
+  aliasedUnion: AliasedUnion,
+  unionedObjects: { prop1: string } | { prop2: string }
+): string {
+  return ""
+}